refactor: 메서드 분리 후 모듈화
